backend/requirements.txt
# SmartAMS Python Dependencies
# Face Recognition, DLib, and API requirements
# Updated for Python 3.14 compatibility

# Core dependencies
Flask>=3.0.0
Flask-CORS>=6.0.0
python-dotenv>=1.0.0
numpy>=2.0.0
Pillow>=10.0.0

# Face Recognition
face-recognition>=1.3.0
face-recognition-models>=0.3.0

# Computer vision
opencv-python-headless>=4.8.0

# DLib for face recognition - ALREADY INSTALLED from build/dlib
# If you need to reinstall it:
# cd build/dlib && pip install -e . && cd ../..
# The precompiled version in build/dlib is available and preferred

# Database & APIs
supabase>=2.4.0
requests>=2.31.0
psycopg2-binary>=2.9.0

# Firebase Admin SDK (token verification & user management)
firebase-admin>=6.0.0

# QR Code & Security
qrcode>=7.4.0
cryptography>=41.0.0

# Security - Password Hashing & Rate Limiting (CRITICAL)
bcrypt>=4.1.0
flask-limiter>=3.5.0
limits>=3.7.0

# WebSocket & Real-time
Flask-SocketIO>=5.3.0
python-socketio>=5.9.0
python-engineio>=4.7.0

# SMS Notifications
twilio>=9.0.0
boto3>=1.26.0

# Advanced Analytics & ML
scikit-learn>=1.3.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn-extra>=0.3.0

# Caching & Session Management
redis>=5.0.0
hiredis>=2.2.0

# Async Task Processing
celery>=5.3.0
kombu>=5.3.0

# Monitoring & Metrics
prometheus-client>=0.18.0

# Utilities
python-multipart>=0.0.6
gunicorn>=21.0.0
gevent>=23.9.0
gevent-websocket>=0.10.1

# Excel & File Processing
openpyxl>=3.1.0
xlrd>=2.0.0

# Optional for debugging
ipython>=8.15.0

# Note: If dlib installation fails, use the pre-compiled version in build/dlib:
# cd build/dlib && pip install -e .
#
# For face-recognition models, use:
git+https://github.com/ageitgey/face_recognition_models
